WIP #41 - Refactor object hierarchy for event bound collections
Issue #41 is all about wrapping generic ISet instances with the event-bound
wrapper system.  Because of this it makes sense to create a new base type for
event-bound collections that requires only the implementation of ICollection<T>
instead of IList<T>.  This base type can then provide functionality for both
lists, sets and any other collections we decide to add later.

* Move functionality to new base type.

* Mark IEventBoundList as obsolete - it brings no value to the API.

* Create new EventBoundCollectionWrapper type that handles ICollection
  instances.