Resolve #58 - Create extension methods for GUIDs as RFC-4122 byte arrays
The rationale for these is that the .NET framework does not comply with RFC-4122
for the format in which is stores GUIDs.  On a little-endian machine, the byte
orders of the data1, data2 and data3 blocks are in the reverse order.  The
spec states that the bytes should ALWAYS be stored in big-endian format, wheras
Microsoft's implementation stores these three blocks in native format.

These extension methods check the endianness of the current machine and (if it
is little-endian) reverses the byte orders of these three data blocks.  This
ensures that the resultant byte array or Guid complies with RFC-4122.