Resolve #54 - Improve the API for identity parsers.
The static parsing methods available on the IdentityParser type are now marked
obsolete.  These are replaced with a new service type: IGenericIdentityParser.

The generic parser may be implemented as any parser which can parse for multiple
types of entity.  The static IdentityParser type now contains a reference to the
AdHocGenericIdentityParser for backwards-compatibility.  This may be removed in
v4.0.0 and up.